Add an undo command to blame windows
Resolves #543.

Support workflow of Gblame, repeated P to find the change you're looking
for, u to go back to the version that introduced it, and CR to inspect
the commit.

Mnemonic: u to undo and go back to blame at previous commit.

I'm not sure how robust this is, but with very light testing it
seems to work.